+++ app/services/accounting/api/v1/accountingapi/courierrapi/response.go
@@ -1,8 +1,8 @@
 package courierapi
 
 type CourierAccountResponse struct {
-	CourierID uint64 `json:"courier_id,omitempty" yaml:"courier_id,omitempty"`
-	Balance    int64  `json:"balance" yaml:"balance"`
+	CourierID int `json:"courier_id,omitempty" yaml:"courier_id,omitempty"`
+	Balance   int `json:"balance" yaml:"balance"`
 }
 
 type ReturnAccountingResponseList struct {
