+++ app/services/accounting/api/v1/accountingapi/restaurantapi/request.go
@@ -1,13 +1,13 @@
 package Restaurantapi
 
 type NewRestaurantAccountRequest struct {
-	RestaurantID uint64 `json:"restaurant_id" yaml:"restaurant_id"`
+	RestaurantID int `json:"restaurant_id" yaml:"restaurant_id"`
 }
 
 type AddRestaurantAccountRequest struct {
-	Amount int64 `json:"amount" yaml:"amount"`
+	Amount int `json:"amount" yaml:"amount"`
 }
 
 type SubBalanceRestaurantAccountRequest struct {
-	Amount int64 `json:"amount" yaml:"amount"`
+	Amount int `json:"amount" yaml:"amount"`
 }
