+++ app/services/accounting/api/v1/accountingapi/restaurantapi/response.go
@@ -1,8 +1,8 @@
 package Restaurantapi
 
 type RestaurantAccountResponse struct {
-	RestaurantID uint64 `json:"restaurant_id,omitempty" yaml:"restaurant_id,omitempty"`
-	Balance      int64  `json:"balance" yaml:"balance"`
+	RestaurantID int `json:"restaurant_id,omitempty" yaml:"restaurant_id,omitempty"`
+	Balance      int `json:"balance" yaml:"balance"`
 }
 
 type ReturnAccountingResponseList struct {
